Clear clippy --all-targets backlog across benches + tests
CI's clippy step runs without --all-targets, which meant benches and
integration-test targets had accumulated ~147 lint warnings + several
hard compile errors over time (criterion 0.8 deprecated black_box,
num-dual/nalgebra API drift, redundant closures, needless range
loops, a stale 3.14 literal flagged as approximating PI, etc.). This
clears the debt so the next commit can tighten CI to fail on it.

Mechanical changes:
- Replace `criterion::black_box` (deprecated) with `std::hint::black_box`
  across all nine bench files (~106 call sites)
- Unwrap redundant closures (`|v| fn(v)` -> `fn`) across benches +
  tests via `cargo clippy --fix` (~73 sites)
- Allow `needless_range_loop` at file level in the 11 scientific
  tests/benches where parallel-indexed `for i in 0..n { ... a[i] ... b[i] ... }`
  is idiomatic and clearer than `.iter().enumerate()`
- Allow `dead_code` at module level in `benches/common/mod.rs` (each
  bench binary only uses a subset of the shared helpers)
- Fix `benches/common/mod.rs::tridiagonal`: the `x[0] - x[0]` pattern
  to construct a zero of the generic `Scalar` type is intentional;
  allow `clippy::eq_op` with explanatory comment
- Fix `benches/comparison.rs`: `black_box(x_dv.clone())` ->
  `black_box(&x_dv)` (num-dual's `gradient` / `jacobian` /
  `hessian` API expects `&DVector<_>`)
- Fix `benches/diffop.rs`: add `use num_traits::Float` to bring
  `.exp()` into scope (pre-existing compile break not previously
  caught because benches aren't clippy'd in CI)
- Fix `tests/forward_mode.rs`: change test value 3.14 -> 2.5 to
  avoid clippy flagging it as approximating PI
- Add `#[allow(clippy::too_many_arguments)]` to
  `check_hypot_jet_non_finite_higher` test helper in
  `tests/gpu_stde.rs` (8 args, deliberate for the GPU kernel
  signature it exercises)

Library fix:
- `src/gpu/mod.rs::compute_chunk_size`: replace manual guarded
  division with `u64::checked_div`

echidna-optim fixes:
- `echidna-optim/tests/sparse_implicit.rs`: `assign_op_pattern`
  auto-fix (`s = s + x` -> `s += x`, 2 sites)
- `echidna-optim/tests/ws4_diagnostics.rs`: use struct-literal
  `LbfgsConfig { memory: 3, convergence: ConvergenceParams { max_iter: 50, ..default() }, ..default() }`
  instead of `cfg.memory = 3; cfg.convergence.max_iter = 50;`
  (Default-initialisation-then-mutate pattern)

All 1024 integration tests pass (889 echidna + 135 echidna-optim).
All 72 GPU tests still green on Metal. `cargo clippy --all-targets
-- -D warnings` now clean on both crates.

+++ tests/laurent.rs
@@ -446,10 +446,10 @@ fn hypot_both_identically_zero() {
     assert_eq!(r.pole_order(), 0);
     for k in 0..4 {
         assert_eq!(
-            r.coeff(k as i32),
+            r.coeff(k),
             0.0,
             "hypot(zero, zero) must stay all-zero; coeff({k}) = {}",
-            r.coeff(k as i32)
+            r.coeff(k)
         );
     }
 }
